pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.131.0",
    "fastapi-mcp>=0.4.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.41.0",
    "watchdog>=6.0.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
